Added TWCC service into separate root folder to separate it from the client folder
